requirements.txt
feedparser==5.2.1
python-dateutil==2.4.2
pytz==2015.6
requests==2.2.1
PyYAML==3.12
